version 0.0.1 Exception Handling

+++ src/main/java/com/example/app/controller/ProductServiceController.java
@@ -2,6 +2,7 @@
 // The Rest Controller class - ProductServiceController.java
 package com.example.app.controller;
 
+import com.example.app.exception.ProductNotfoundException;
 import com.example.app.model.Product;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -42,6 +43,8 @@ public class ProductServiceController {
     @RequestMapping(value = "/products/{id}",method = RequestMethod.PUT)
     public ResponseEntity<Object> updateProduct(@PathVariable("id") String id,@RequestBody Product product)
     {
+        if(!productRepo.containsKey(id))
+            throw new ProductNotfoundException();
         productRepo.remove(id);
         product.setId(id);
         productRepo.put(id,product);
@@ -50,6 +53,8 @@ public class ProductServiceController {
     // suppression d'un produit existant
     @RequestMapping(value = "/products/{id}",method = RequestMethod.DELETE)
     public ResponseEntity<Object> delete(@PathVariable("id") String id){
+        if(!productRepo.containsKey(id))
+            throw new ProductNotfoundException();
         productRepo.remove(id);
         return new ResponseEntity<>("Product is deleted successfully",HttpStatus.NO_CONTENT);
     }
